refactor: revamp reporters & avoid ref to async tx
